Allow viewing an email by userId
Before viewing an email required specifying the email for the user of
the message. Using the user's `userId` wasn't an option. This blocked
querying the viewInBrowser endpoint for userID-based projects.

Similar to other `ApiResource` methods, this changes `view` to make
email optional in favor of specifying the user's `userId` in `attrs`
which send along with the request as query parameters.