Permit nil email for updating by user ID
Updating a user or a user's subscriptions using `Iterable::User#update`
and `Iterable::User#update_subscriptions`, respectively, using the user
ID requires passing `nil` as the first argument to each. This avoid the
error "invalid email", but conflicts with Sorbet's type definitions. To
resolve the conflict and maintain functionality permit passing `nil` to
each of the methods.